refactor: shard public forms router (#1669)
* refactor(public-forms): moved subroutes out of main publicforms router

realised that public form is getting bloated adn we should not keep every route in the same file but
rather, shard by responsiblity. hence, this (mini) pr is for doign that before starting on auth
router to prevent excess bloat from accumulating

* refactor(public-form/tests): shards tests out also

* style(api): changed folder name to forms instead of public
