+++ src/app/routes/api/v3/forms/public-forms.routes.ts
@@ -1,0 +1,11 @@
+import { Router } from 'express'
+
+import { PublicFormsFeedbackRouter } from './public-forms.feedback.routes'
+import { PublicFormsFormRouter } from './public-forms.form.routes'
+import { PublicFormsSubmissionsRouter } from './public-forms.submissions.routes'
+
+export const PublicFormsRouter = Router()
+
+PublicFormsRouter.use(PublicFormsSubmissionsRouter)
+PublicFormsRouter.use(PublicFormsFeedbackRouter)
+PublicFormsRouter.use(PublicFormsFormRouter)
